+++ packages/ui/resolution/resolution.js
@@ -145,10 +145,10 @@ function getScreenType (rez) {
 		rez.height = swap;
 	}
 
-	// Loop through resolutions, last->first, largest->smallest
-	for (let i = types.length - 1; i >= 0; i--) {
-		// Does the current resolution fit inside this screenType definition? If so, save it as the current best match.
-		if (rez.height <= types[i].height && rez.width <= types[i].width) {
+	// Loop through resolutions, first->last, smallest->largest
+	for (let i = 0; i <= types.length - 1; i++) {
+		// Does this screenType definition fit inside the current resolution? If so, save it as the current best match.
+		if (rez.height >= types[i].height && rez.width >= types[i].width) {
 			bestMatch = types[i].name;
 		}
 	}
@@ -189,6 +189,9 @@ function calculateFontSize (type) {
 		size = scrObj.height / scrObj.width * scrObj.pxPerRem;
 	} else {
 		size = scrObj.pxPerRem;
+		if (workspaceBounds.width > scrObj.width && workspaceBounds.height > scrObj.height ) {
+			size = parseInt(workspaceBounds.height * scrObj.pxPerRem / scrObj.height);
+		}
 	}
 	return size + 'px';
 }