+++ packages/ui/CHANGELOG.md
@@ -8,6 +8,11 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 - `ui/Layout.Cell` prop `grow` to expand its size to the container
 
+### Changed
+
+- `ui/resolution` to select screenType which is smaller than actual screen size
+- `ui/resolution` to calculate the base font size by considering screenType and actual screen size
+
 ### Fixed
 
 - `ui/ViewManager` to set index prop properly when reverseTransition prop is given